Created main.py algorithm to test who would win in a 10,000 experiment simulation with both a set addition amount (7) or a random amount (1-7).
I could be right, but currently I think this is bugged, because it returns that only student 1 wins, whether they go first or last, whether the number is set at 7 or is random, student 1 always wins. I'll take a detailed look through the debugger while running to see if anything has been missed.

I'm suprised how easily this was to code and about how focused I was able to be, its 2am but I've slept all day. Started this in just class today and was able to flesh out a pretty good starting point, with good formatting.